DAS-2067 - Migrate formatting.py.
