+++ xarray/core/datatree.py
@@ -23,6 +23,7 @@ from xarray.core.datatree_mapping import (
     check_isomorphic,
     map_over_subtree,
 )
+from xarray.core.formatting import datatree_repr
 from xarray.core.formatting_html import (
     datatree_repr as datatree_repr_html,
 )
@@ -40,7 +41,6 @@ from xarray.core.utils import (
 )
 from xarray.core.variable import Variable
 from xarray.datatree_.datatree.common import TreeAttrAccessMixin
-from xarray.datatree_.datatree.formatting import datatree_repr
 from xarray.datatree_.datatree.ops import (
     DataTreeArithmeticMixin,
     MappedDatasetMethodsMixin,
